Relative placement is slowly getting done

+++ scoring/src/main/java/nz/geek/goodwin/wsdc/domain/InterimResult.java
@@ -1,10 +1,11 @@
 package nz.geek.goodwin.wsdc.domain;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 
 /**
  * @author thomas.goodwin
  */
-public record InterimResult<T>(Person dancer, List<Score<T>> scores, BigDecimal headJudgeScore, BigDecimal totalScore) {
+public record InterimResult<T>(Person dancer, List<Score<T>> scores, BigInteger headJudgeScore, BigDecimal totalScore) {
 }
